Switch from service httpd to apachectl for Apache
The command 'service httpd fullstatus' is incompatible with Debian and
any other distribution (and some BSDs as well?) that names the Apache
service anything other than httpd. In Debian, for example, the service is
named apache2.

The apachectl command is the best option for cross-distro usage, and
avoids issues potentially arising in the future as distros migrate from
using 'service' to using 'systemctl' for systemd.

These changes have been tested successfully on CentOS 5 (previously
working) and Debian 6 (not previously working) servers.