multiple makefile targets
In RPM spec files, putting doc files in place is handled by the %doc and
%license macros.  In other words, we need to be able to explictly install
everything _but_ the docs.  This change preserves the existing behavior of
`make install`/`make uninstall`, but splits the actions out into separate
targets.  This will allow us to only run `make install-base` and `make
install-man` in the spec file.  Additionally, this change templatizes the
recap.cron file and generates it appropriately.